Use system file selector. Pass URIs to Keepass Droid.
Removed dependency on external file chooser in favour of the system file chooser.

KP Droid now supports URIs in Intents, when previously it only supported files. This is great news because Android content providers communicate using URIs, so we can now pass the locations of files which aren't on the device (such as Google Drive files) to KP Droid.