Support for KPNFC smartcard.
This adds support for JavaCard 2.1-compatible contactless smartcards running the KeePassNFC applet. If the NFC tag contains the applet, then the encryption key will be written to the smartcard securely. If the NFC tag does not run the applet (e.g. it's a simple read/write NDEF tag) then everything works as previously, i.e. the encryption key will be written to the tag in the clear.