Switch back to ACTION_OPEN_DOCUMENT.
It seems that Google / application providers still aren't getting this right. ACTION_GET_CONTENT is to be used for immediate access to content -- it's not intended for long-term access (and in fact Google Drive won't give you a persistable grant when you use ACTION_GET_CONTENT). ACTION_OPEN_DOCUMENT is for longer-term access to the data.

 Unfortunately most content providers don't implement the Storage Access Framework and thus don't support ACTION_OPEN_DOCUMENT. Dropbox, for example, only supports ACTION_GET_CONTENT. The URI thus returned happens to be accessible across reboots, but this isn't a guarantee and, more importantly, is not supported by Google Drive (there appears to be no way to get persistent access to a URI returned by ACTION_GET_CONTENT from Google Drive).

  So for the moment I'm using ACTION_OPEN_DOCUMENT and waiting for content provider authors to support the SAF.